fix(qwik): 🐛 Add support for TypeScript types to the Qwik SDK
