+++ packages/sdks/output/qwik/.gitignore
@@ -1,0 +1,3 @@
+types
+dist
+lib
