+++ packages/sdks/output/qwik/.npmignore
@@ -1,0 +1,6 @@
+src
+dist
+.yarn
+vite.config.ts
+tsconfig.json
+DEVELOPER.md
