+++ packages/sdks/output/qwik/package.json
@@ -1,12 +1,12 @@
 {
   "name": "@builder.io/sdk-qwik",
-  "version": "0.0.15",
+  "version": "0.0.17",
   "description": "Builder.io Qwik SDK",
   "type": "module",
   "main": "./lib/index.qwik.cjs",
   "module": "./lib/index.qwik.mjs",
   "qwik": "./lib/index.qwik.mjs",
-  "types": "./types.d.ts",
+  "types": "./types/index.d.ts",
   "exports": {
     ".": {
       "import": "./lib/index.qwik.mjs",
@@ -14,16 +14,17 @@
     }
   },
   "scripts": {
-    "build.lib": "vite build --mode lib",
+    "typecheck": "tsc --noEmit",
+    "build.lib": "tsc && vite build --mode lib",
     "release:dev": "npm version prerelease --no-git-tag-version && npm publish --tag latest --access public"
   },
   "devDependencies": {
-    "@builder.io/qwik": "0.0.103",
+    "@builder.io/qwik": "0.0.108",
     "@types/node": "latest",
     "typescript": "4.7.4",
     "vite": "3.0.2"
   },
   "peerDependencies": {
-    "@builder.io/qwik": "0.0.103"
+    "@builder.io/qwik": "0.0.108"
   }
 }
