--- packages/sdks/output/qwik/types.d.ts
@@ -1,8 +1,0 @@
-declare module '@builder.io/sdk-qwik' {
-  const getContent: (options: {
-    model: string;
-    apiKey: string;
-    userAttributes: Record<string, string>;
-  }) => Promise<any>;
-  const RenderContent: any;
-}
