+++ packages/sdks/output/qwik/vite.config.ts
@@ -6,7 +6,7 @@ export default defineConfig(() => {
     build: {
       target: 'es2020',
       lib: {
-        entry: './src/index.js',
+        entry: './src/index.ts',
         formats: ['es', 'cjs'],
         fileName: (format) => `index.qwik.${format === 'es' ? 'mjs' : 'cjs'}`,
       },
