+++ packages/sdks/overrides/qwik/src/constants/target.ts
@@ -1,2 +1,2 @@
 /** This file should be overriden for each framework. Ideally this would be implemented in Mitosis.  */
-export const TARGET = 'qwik';
+export const TARGET = 'qwik' as any;
