+++ packages/sdks/src/components/render-block/render-repeated-block.lite.tsx
@@ -1,7 +1,6 @@
 import { setContext } from '@builder.io/mitosis';
-import BuilderContext, {
-  BuilderContextInterface,
-} from '../../context/builder.context.lite';
+import BuilderContext from '../../context/builder.context.lite';
+import type { BuilderContextInterface } from '../../context/types.js';
 import type { BuilderBlock } from '../../types/builder-block';
 import RenderBlock from './render-block.lite';
