+++ packages/sdks/src/constants/builder-registered-components.ts
@@ -14,7 +14,7 @@ import { componentInfo as textComponentInfo } from '../blocks/text/component-inf
 import { default as Text } from '../blocks/text/text.lite';
 import { componentInfo as videoComponentInfo } from '../blocks/video/component-info.js';
 import { default as Video } from '../blocks/video/video.lite';
-import type { RegisteredComponent } from '../context/builder.context.lite';
+import type { RegisteredComponent } from '../context/types.js';
 import { componentInfo as embedComponentInfo } from '../blocks/embed/component-info.js';
 import { default as embed } from '../blocks/embed/embed.lite';
 import { default as Img } from '../blocks/img/img.lite';
