+++ packages/sdks/src/functions/get-block-actions.ts
@@ -1,4 +1,4 @@
-import type { BuilderContextInterface } from '../context/builder.context.lite';
+import type { BuilderContextInterface } from '../context/types.js';
 import { BuilderBlock } from '../types/builder-block.js';
 import { evaluate } from './evaluate.js';
 import { getEventHandlerName } from './event-handler-name.js';
