+++ packages/sdks/src/functions/get-processed-block.ts
@@ -1,4 +1,4 @@
-import type { BuilderContextInterface } from '../context/builder.context.lite';
+import type { BuilderContextInterface } from '../context/types.js';
 import { BuilderBlock } from '../types/builder-block.js';
 import { evaluate } from './evaluate.js';
 import { set } from './set.js';
